server/linebot/requirements.txt
flask
requests
pandas
line-bot-sdk
PyPDF2
python-dotenv
PyCryptodome